Restructure top-level packages
