+++ src/talondoc/__init__.py
@@ -4,8 +4,8 @@ from typing import Optional
 import click
 
 from ._version import __version__
+from .analysis.dynamic import cache_builtin
 from .autogen import autogen
-from .cache_builtin import cache_builtin
 
 
 @click.group(name="talondoc")
