+++ src/talondoc/_util/logging.py
@@ -1,11 +1,12 @@
 import logging
 import sys
+from typing import cast
 
 
 def getLogger(name: str) -> logging.LoggerAdapter:
     if "sphinx" in sys.modules:
         import sphinx.util.logging
 
-        return sphinx.util.logging.getLogger(name)  # type: ignore
+        return cast(logging.LoggerAdapter, sphinx.util.logging.getLogger(name))
     else:
         return logging.LoggerAdapter(logger=logging.getLogger(name), extra={})
