+++ src/talondoc/analysis/static/__init__.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 from typing import Optional
 
-from .. import Registry
-from ..entries import Package
-from ..entries.abc import Location
+from ..registry import Registry
+from ..registry.entries import Package
+from ..registry.entries.abc import Location
 from .python import analyse_files as analyse_python_files
 from .talon import analyse_files as analyse_talon_files
