+++ src/talondoc/sphinx/_util/typing.py
