+++ src/talondoc/sphinx/directives/command/__init__.py
@@ -1,13 +1,12 @@
-import re
 import sys
-from typing import Iterator, List, Optional, Tuple
+from typing import List
 
 from sphinx import addnodes
 from sphinx.util.typing import OptionSpec
 from typing_extensions import override
 
 from ...._util.logging import getLogger
-from ...util.typing import flag, optional_str, optional_strlist
+from ..._util.typing import flag, optional_strlist
 from ..errors import AmbiguousSignature, UnmatchedSignature
 from .abc import TalonDocCommandDescription
