--- src/talondoc/sphinx/directives/util.py
@@ -1,10 +1,0 @@
-import re
-from typing import Iterator, Sequence, Union
-
-from ...registry import Registry
-from ...registry import entries as talon
-
-
-def wildcard_pattern(pattern: str) -> re.Pattern:
-    """Compile a pattern with wildcards to a regular expression."""
-    return re.compile(".*".join(map(re.escape, pattern.split("*"))))
