+++ src/talondoc/sphinx/domains.py
@@ -2,7 +2,7 @@ from sphinx.domains import Domain
 from sphinx.environment import BuildEnvironment
 
 from .._util.logging import getLogger
-from ..registry import Registry
+from ..analysis.registry import Registry
 from .directives.command import TalonCommandDirective
 from .directives.command.hlist import TalonCommandHListDirective
 from .directives.command.table import TalonCommandTableDirective
